requirements.txt
python-docx
python-dotenv
openai
azure-identity